refactor: simplify codebase structure (fixes #111)
- Remove dead knowledge/ingest.py and its tests (zero production imports)
- Clean knowledge/__init__.py and platform/__init__.py barrel re-exports
- Inline serial.py into engine.py, move invoke_runner to state.py
- Move infra/logging.py to core/logging.py, dissolve infra/ package
- Merge session/context.py into session/prompt.py
